+++ idea/testData/refactoring/pullUp/k2k/removeVisibilityOnOverride.kt
@@ -1,0 +1,8 @@
+open class A {
+
+}
+
+open class <caret>B : A() {
+    // INFO: {"checked": "true", "toAbstract": "true"}
+    protected fun foo() {}
+}